split sql test
